Add memory_range.go which defines the interface type Range and implements it.
Range is a bounded, or unbounded above, range of memory sizes (*MemSize). In
this implementation the lower and upper sizes may be negative, but the range
may only be unbounded above.

Operations of scaling, containment, string representation and construction,
equality, IsBounded(), and Floor() and Ceiling() are implemented. A full suite
of tests is included and each method and exported function is documented.

[#93189380]