Add memory_bucket.go which defines the Bucket and StackBucket interfaces.
A Bucket is a range of supported memory sizes together with a weight used to
determine relative proportions of the total memory to use. Buckets are used to
help calculate memory allocations for a JRE based upon an overall memory
allocation and optional constraints. StackBucket is a specialisation of
Bucket which also knows an individual stack allocation default.

[#93189380]