Produce non-zero exit code and message if any allocation is zero.
After a successful balancing of the memory into buckets, check that none of
the explicit memory allocations is zero (after rounding down to whole
kilobytes); if one is zero, fail with an error message. Code improved to make
this easier to read in the Balance() method and removed the dependency upon
internal state for tracking the error in subsequent method calls.

The tests are improved to increase coverage and make failure testing clearer.

[#94164248]