Report an error if -totMemory omitted, or specifies less than 1K.
Previously, if the -totMemory flag was omitted the memory calculator would
allocate the minimum of the size ranges supplied in -memorySizes. However, if
some memory types had a zero lower-bound (the default), the allocation would
be zero. This is never a valid memory option.

This change makes an omitted -totMemory flag, or a flag that specifies less
than one kilobyte, an error, resulting in a non-zero exit code (and a message
on stderr).

Some code in the memory calculator is removed.

[#94311618]