Issue correct error message when allocation results in zero bytes.
Previously when allocation results in zero bytes for a particular memory type
the memory calculator issues a message which is garbled. This commit fixes the
format of this message, and adds a test to check it is correctly produced.

The Go compiler version is updated to 1.5.1, and the dependencies for running
the tests have changed. Finally, defaults on flags are reported differently in
the Go 1.5.1 flag package, so the standalone flags tests have changed
accordingly.