Support Java 7
Introduce a -poolType flag which can take the value 'metaspace' for Java 8 and
later or 'permgen' for Java 7. Use a new set of estimation functions, based on
measurements of sample applications running on Java 7, for permgen poolType.

Avoid making the new flag Java version specific because that would require
maintenance of the memory calculator for each new Java version regardless of
whether or not memory calculation changes are needed.

[#138498389]