Don't Print JVM Memory Configuration on Startup
Previously, the JVM memory configuration was printed on startup.  Because the
memory calculator stdout is captured, this printing had to be done to stderr.
This caused some alarm (mostly by Sabby), so an alternative was created. This
change removes the printing from the memory calculator, instead requiring the
buildpack to print it.