Version 4
This change makes massive, incompatible updates to the project resulting in a
bump to version 4.  The code has been reworked to be more modular and simpler.
It takes lessons from the current building of Cloud Native Buildpacks to
become more idiomatic go.  Finally, the input flags are now POSIX-style rather
than Go style.